Refactored the Server struct.
Added http field to holde the server instance.
