+++ internal/server/server_test.go
@@ -1,4 +1,4 @@
-package api
+package server
 
 import (
 	"net/http"
